+++ lib/flat/web.d.ts
@@ -1,5 +1,0 @@
-import type { Linter } from "eslint";
-declare type Config = Linter.Config[];
-
-export let config: Config;
-export default config;