+++ src/extra/prettier.ts
@@ -1,0 +1,6 @@
+import prettierPlugin from "eslint-plugin-prettier/recommended";
+
+import * as utils from "../env.js";
+
+export default utils.usingPrettier() ? [] : [prettierPlugin];
+