+++ src/legacy.ts
@@ -1,0 +1,1 @@
+throw Error("Legacy config is not support now.");